requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
celery
redis
openpyxl
gunicorn
